Split files and add logging
